Deleted Death From Front Headshot.anim; Fixed HealthBar's upding health; Added IShootRayCalculator.cs interface, ShootRayCalculatorWithCamera.cs and ShootRayCalculatorWithoutCamera.cs to for different realizations of the ray; Now Weapon.cs uses a IShootRayCalculator.

+++ Assets/Code/Scripts/UI/HealthBar/HealthBar.cs
@@ -31,7 +31,7 @@ namespace FPS.UI
 
         private void HandleHealthChanging(int newValue, int defaultValue)
         {
-            _progress.fillAmount = newValue / defaultValue;
+            _progress.fillAmount = (float)newValue / (float)defaultValue;
         }
 
         #endregion
